chore: create views folder and move not found component
